luna_v.3.3.3/luna-index-manager/requirements.txt
requests==2.18.4
tornado==5.0.1
logbook==1.1.0
ujson==1.35
jsonschema==2.6.0
sqlalchemy==1.2.5
cx-Oracle==6.3.1
psycopg2==2.7.1
alembic==1.0.0
six==1.11.0
python-dateutil==2.7.3
pytz==2018.5
extras/luna3-v.0.0.50.tar.gz
